set up example
