+++ examples/camera-access/index.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { App } from './app.js'
+import { App } from './src/app.js'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
